+++ controler/ConnecteurControler.php
@@ -64,6 +64,7 @@ class ConnecteurControler extends PastellControler
         $this->setCanEditConnector($id_e);
         $this->setDroitImportExportConfig($id_e);
         $this->setDroitLectureOnUtilisateur($id_e);
+        $this->setDroitsDaemon($id_e);
     }
 
     public function hasDroitEdition($id_e): void
@@ -735,6 +736,7 @@ class ConnecteurControler extends PastellControler
         $connecteur_info = $this->getConnecteurEntiteSQL()->getInfo($id_ce);
         $id_e = $connecteur_info['id_e'];
         $this->setDroitLectureOnUtilisateur($id_e);
+        $this->setDroitsDaemon($id_e);
         $this->verifDroitOnConnecteur($id_ce);
 
         $documentType = ($connecteur_info['global']) ?