+++ controler/EntiteControler.php
@@ -1,6 +1,7 @@
 <?php
 
 use Pastell\Service\Crypto;
+use Pastell\Service\Droit\DroitService;
 use Pastell\Service\Entite\EntityCreationService;
 use Pastell\Service\Entite\EntityUpdateService;
 use Pastell\Service\FeatureToggleService;
@@ -34,6 +35,7 @@ class EntiteControler extends PastellControler
                 ->isEnabled(CDGFeature::class)
         );
         $this->setDroitLectureOnUtilisateur($id_e);
+        $this->setDroitsDaemon($id_e);
     }
 
     private function getAgentSQL()
@@ -533,10 +535,18 @@ class EntiteControler extends PastellControler
         if ($this->getConnecteurEntiteSQL()->getAll($id_e)) {
             return false;
         }
+        if ($this->getDaemonSQL()->getDaemonByEntity($id_e) !== null) {
+            return false;
+        }
         return true;
     }
 
-    public function supprimerAction()
+    /**
+     * @throws UnrecoverableException
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function supprimerAction(): void
     {
         $recuperateur = new Recuperateur($_GET);
         $id_e = $recuperateur->getInt('id_e', 0);
@@ -552,7 +562,7 @@ class EntiteControler extends PastellControler
             Journal::MODIFICATION_ENTITE,
             $info['entite_mere'],
             $this->getId_u(),
-            "Suppression",
+            'Suppression',
             "Suppression de l'entité $id_e qui contenait : \n" . implode("\n,", $info)
         );
         $this->getEntiteSQL()->delete($id_e);
@@ -561,7 +571,12 @@ class EntiteControler extends PastellControler
         $this->redirect("/Entite/detail?id_e={$info['entite_mere']}");
     }
 
-    public function activerAction()
+    /**
+     * @throws UnrecoverableException
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function activerAction(): void
     {
         $recuperateur = new Recuperateur($_GET);
         $id_e = $recuperateur->getInt('id_e', 0);
@@ -827,5 +842,159 @@ class EntiteControler extends PastellControler
         $lastErrors = $importConfigService->getLastErrors();
         $this->setLastMessage('Les données ont été importées<br/>' . implode('<br/>', $lastErrors));
         $this->redirect("/Entite/detail?id_e=$id_e");
+    }
+
+
+    /**
+     * @throws NotFoundException
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function daemonAction(): void
+    {
+        $this->daemonData();
+        $this->setViewParameter('page_url', 'index');
+        $this->setViewParameter('twigTemplate', 'daemon/entity/index.html.twig');
+        $this->setViewParameter('page_title', 'Gestionnaire de tâches local');
+        $this->renderDefault();
+    }
+
+
+    /**
+     * @throws LastMessageException
+     * @throws LastErrorException
+     * @throws NotFoundException
+     */
+    public function daemonContentAction(): void
+    {
+        $this->daemonData();
+        $this->setViewParameter('twigTemplate', 'daemon/entity/index_content.html.twig');
+        header('Content-type: text/html; charset=utf-8;');
+        $this->renderDefault();
+    }
+
+    /**
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    private function daemonData(): void
+    {
+        $recuperateur = $this->getGetInfo();
+        $id_e = $recuperateur->getInt('id_e');
+        $this->verifDroit(
+            $id_e,
+            DroitService::getDroitLecture(DroitService::DROIT_DAEMON)
+        );
+        $this->setViewParameter('id_e', $id_e);
+        $this->setDroitsDaemon($id_e);
+
+        $daemon = $id_e === EntiteSQL::ID_E_ENTITE_RACINE ?
+            $this->getDaemonSQL()->getGlobalDaemon() :
+            $this->getDaemonSQL()->getDaemonByEntity($id_e);
+
+        if ($daemon === null) {
+            $this->setLastError('Aucun gestionnaire de tâche lié à cette entité.');
+            $this->redirect("Entite/detail?id_e=1$id_e");
+        } else {
+            $this->setViewParameter('menu_gauche_select', 'Entite/daemon');
+            $this->setViewParameter('nb_worker_actif', $this->getWorkerSQL()->getNbActifForDaemon($daemon->id_daemon));
+            $this->setViewParameter('job_stat_info', $this->getJobQueueSQL()->getStatInfoForDaemon($daemon->id_daemon));
+            $this->setViewParameter('sub_title', 'Liste de tous les travaux');
+            $this->setViewParameter('return_url', "Entite/daemon?id_e=$daemon->id_e");
+            $job_list = $this->getJobQueueSQL()->getJobsByDaemon($daemon->id_daemon, 20, 0);
+            $this->setViewParameter('job_list', $job_list);
+            $this->setViewParameter('daemon', $daemon);
+        }
+    }
+
+    /**
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function daemonUnlockAllAction(): void
+    {
+        $recuperateur = $this->getGetInfo();
+        $id_e = $recuperateur->getInt('id_e');
+        $this->verifDroit(
+            $id_e,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
+
+        $daemon = $id_e === EntiteSQL::ID_E_ENTITE_RACINE ?
+            $this->getDaemonSQL()->getGlobalDaemon() :
+            $this->getDaemonSQL()->getDaemonByEntity($id_e);
+        if ($daemon === null) {
+            $this->setLastError('Aucun gestionnaire de tâche lié à cette entité.');
+            $this->redirect("Entite/detail?id_e=$id_e");
+        } else {
+            $this->getWorkerSQL()->menageAll();
+            $this->getJobQueueSQL()->unlockAll($daemon->id_daemon);
+        }
+
+        $this->redirect('Entite/daemon?id_e=' . $id_e);
+    }
+
+    /**
+     * @throws LastErrorException
+     * @throws LastMessageException
+     * @throws NotFoundException
+     * @throws UnrecoverableException
+     */
+    public function jobAction(): void
+    {
+        $recuperateur = $this->getGetInfo();
+        $id_e = $recuperateur->getInt('id_e');
+        $daemon = $id_e === EntiteSQL::ID_E_ENTITE_RACINE ?
+            $this->getDaemonSQL()->getGlobalDaemon() :
+            $this->getDaemonSQL()->getDaemonByEntity($id_e);
+        if ($daemon === null) {
+            $this->setLastError('Aucun gestionnaire de tâche lié à cette entité.');
+            $this->redirect("Entite/detail?id_e=$id_e");
+        } else {
+            $this->setViewParameter('id_e', $id_e);
+            $this->setViewParameter('menu_gauche_select', 'Entite/job');
+
+            $this->verifDroit($id_e, DroitService::getDroitLecture(DroitService::DROIT_DAEMON));
+            $this->setViewParameter('twigTemplate', 'daemon/entity/job.html.twig');
+            $this->setViewParameter('page_title', 'Gestionnaire de tâches local');
+            $filtre = $recuperateur->get('filtre', '');
+            if ($filtre) {
+                $this->setViewParameter('page_url', "job?filtre=$filtre");
+                $this->setViewParameter('menu_gauche_select', "Entite/job?filtre=$filtre");
+            } else {
+                $this->setViewParameter('page_url', 'job');
+            }
+
+            $sub_title_array = [
+                'actif' => 'Liste des travaux actifs',
+                'lock' => 'Liste des travaux suspendus',
+                'wait' => 'Liste des travaux en retard',
+            ];
+
+            $this->setViewParameter('sub_title', $sub_title_array[$filtre] ?? 'Liste de tous les travaux');
+            $this->setViewParameter('unlock_all_action', 'app.legacy.entite_daemonUnlockAll');
+
+            $this->setViewParameter('offset', $recuperateur->getInt('offset', 0));
+            $this->setViewParameter('limit', 50);
+            $this->setViewParameter('filtre', $filtre);
+            $this->setViewParameter('id_e', $id_e);
+
+            $this->setViewParameter(
+                'return_url',
+                "Entite/job?filtre=$filtre&offset=" . $this->getViewParameterByKey('offset') . "&id_e=$id_e"
+            );
+
+            $this->setViewParameter('count', $this->getJobQueueSQL()->getNbJob($filtre, $daemon->id_daemon));
+            $this->setViewParameter(
+                'job_list',
+                $this->getJobQueueSQL()->getFilteredJobList(
+                    $this->getViewParameterByKey('limit'),
+                    $this->getViewParameterByKey('offset'),
+                    $filtre,
+                    $daemon->id_daemon
+                )
+            );
+        }
+        $this->renderDefault();
     }
 }