+++ model/WorkerSQL.php
@@ -152,10 +152,19 @@ class WorkerSQL extends SQL
         return $this->queryOneCol($sql);
     }
 
-    public function getNbActif()
+    public function getNbActif(): int
     {
-        $sql = "SELECT count(*) FROM worker WHERE termine=0";
+        $sql = 'SELECT count(*) FROM worker WHERE termine=0';
         return $this->queryOne($sql);
+    }
+
+    public function getNbActifForDaemon(int $id_daemon): int
+    {
+        $sql = 'SELECT count(*) 
+            FROM worker w 
+            JOIN job_queue jq ON jq.id_job = w.id_job 
+            WHERE termine=0 AND jq.id_daemon = ?';
+        return $this->queryOne($sql, $id_daemon);
     }
 
     public function getActif($offset = 0, $limit = 20)
@@ -180,27 +189,6 @@ class WorkerSQL extends SQL
     {
         $sql = "DELETE FROM worker WHERE termine=1";
         $this->query($sql);
-    }
-
-
-    public function getNbJob($filtre)
-    {
-        $sql = "SELECT count(*) " .
-                " FROM job_queue" .
-                " LEFT JOIN worker ON job_queue.id_job = worker.id_job " .
-                " WHERE 1=1 ";
-
-        if ($filtre === 'lock') {
-            $sql .= ' AND job_queue.is_lock=1';
-        }
-        if ($filtre === 'wait') {
-            $sql .= ' AND job_queue.next_try < NOW()';
-        }
-        if ($filtre === 'actif') {
-            $sql .= ' AND worker.termine = 0';
-        }
-
-        return $this->queryOne($sql);
     }
 
     public function getJobListWithWorkerForConnecteur($id_ce)