+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Suppression de la fonction WorkerSQL::getJobToLaunch remplacé par WorkerSQL::getJobsToLaunch #2237
 - Suppression de la fonction WorkerSQL::getRunningWorkerInfo remplacé par WorkerSQL::getRunningWorker #2237
 - Suppression de la fonction WorkerSQL::getJobListWithWorker remplacé par JobQueueSQL::getFilteredJobList #2233
+- Suppression de la fonction WorkerSQL::getNbJob remplacé par JobQueueSQL::getNbJob #2239
 
 ## Suppression d'éléments dépréciés